Fixed ui validation on mac. Split identification into separate file.
